Tweaked the naming of logcat functions.
